requirements.txt
streamlit
torch
torchvision
transformers
accelerate
Pillow
numpy
opencv-python
